+++ packages/wallet-connectors/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Build as ES module to facilitate tree-shaking.
+
 ## [0.3.2] - 2023-08-17
 
 ### Changed