+++ packages/wallet-connectors/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-11-13
+
 ### Changed
 
 -   (breaking) Make compatible with `@concordium/web-sdk@^7`, which is now a peer dependency.